# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Allied School Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export type LoginResultRole = typeof LoginResultRole[keyof typeof LoginResultRole];


export const LoginResultRole = {
  admin: 'admin',
  director: 'director',
} as const;

export interface LoginResult {
  success: boolean;
  role: LoginResultRole;
  username: string;
}

export interface Student {
  id: number;
  rollNumber: string;
  name: string;
  class: string;
  /** @nullable */
  section?: string | null;
  fatherName: string;
  /** @nullable */
  motherName?: string | null;
  phone: string;
  /** @nullable */
  fatherPhone?: string | null;
  /** @nullable */
  homePhone?: string | null;
  /** @nullable */
  whatsappNumber?: string | null;
  hasWhatsapp?: boolean;
  /** @nullable */
  address?: string | null;
  admissionDate: string;
  monthlyFee: number;
  isActive: boolean;
  /** @nullable */
  leftDate?: string | null;
}

export interface FeeRecord {
  id: number;
  studentId: number;
  /** @nullable */
  studentName?: string | null;
  /** @nullable */
  rollNumber?: string | null;
  /** @nullable */
  class?: string | null;
  month: number;
  year: number;
  amount: number;
  paid: boolean;
  /** @nullable */
  paidDate?: string | null;
}

export interface AttendanceSummary {
  totalDays: number;
  presentDays: number;
  absentDays: number;
  percentage: number;
}

export interface StudentDetail {
  id: number;
  rollNumber: string;
  name: string;
  class: string;
  /** @nullable */
  section?: string | null;
  fatherName: string;
  /** @nullable */
  motherName?: string | null;
  phone: string;
  /** @nullable */
  fatherPhone?: string | null;
  /** @nullable */
  homePhone?: string | null;
  /** @nullable */
  whatsappNumber?: string | null;
  hasWhatsapp?: boolean;
  /** @nullable */
  address?: string | null;
  admissionDate: string;
  monthlyFee: number;
  isActive: boolean;
  /** @nullable */
  leftDate?: string | null;
  feeHistory: FeeRecord[];
  attendanceSummary: AttendanceSummary;
}

export interface StudentInput {
  rollNumber: string;
  name: string;
  class: string;
  section?: string;
  fatherName: string;
  motherName?: string;
  phone: string;
  fatherPhone?: string;
  homePhone?: string;
  whatsappNumber?: string;
  hasWhatsapp?: boolean;
  address?: string;
  admissionDate: string;
  monthlyFee: number;
}

export interface StudentUpdate {
  name?: string;
  class?: string;
  section?: string;
  fatherName?: string;
  motherName?: string;
  phone?: string;
  fatherPhone?: string;
  homePhone?: string;
  whatsappNumber?: string;
  hasWhatsapp?: boolean;
  address?: string;
  monthlyFee?: number;
  isActive?: boolean;
  leftDate?: string;
}

export interface FeeRecordInput {
  studentId: number;
  month: number;
  year: number;
  amount: number;
}

export interface FeeRecordUpdate {
  paid?: boolean;
  paidDate?: string;
}

export interface GenerateMonthlyFeesInput {
  month: number;
  year: number;
}

export interface AttendanceRecord {
  id: number;
  studentId: number;
  /** @nullable */
  studentName?: string | null;
  /** @nullable */
  rollNumber?: string | null;
  /** @nullable */
  class?: string | null;
  date: string;
  present: boolean;
}

export interface AttendanceInput {
  studentId: number;
  date: string;
  present: boolean;
}

export interface AttendanceUpdate {
  present?: boolean;
}

export type AnnouncementTargetRole = typeof AnnouncementTargetRole[keyof typeof AnnouncementTargetRole];


export const AnnouncementTargetRole = {
  all: 'all',
  admin: 'admin',
  parents: 'parents',
} as const;

export interface Announcement {
  id: number;
  title: string;
  content: string;
  targetRole: AnnouncementTargetRole;
  createdAt: string;
}

export type AnnouncementInputTargetRole = typeof AnnouncementInputTargetRole[keyof typeof AnnouncementInputTargetRole];


export const AnnouncementInputTargetRole = {
  all: 'all',
  admin: 'admin',
  parents: 'parents',
} as const;

export interface AnnouncementInput {
  title: string;
  content: string;
  targetRole: AnnouncementInputTargetRole;
}

export interface DashboardStats {
  totalStudents: number;
  activeStudents: number;
  currentMonthFeeExpected: number;
  currentMonthFeeCollected: number;
  currentMonthFeePending: number;
  todayAttendancePresent: number;
  todayAttendanceAbsent: number;
  totalAnnouncements: number;
}

export interface DailyCollection {
  date: string;
  totalCollected: number;
  paymentCount: number;
  payments: FeeRecord[];
}

export interface ClassSummary {
  class: string;
  totalStudents: number;
  feesPaid: number;
  feesPending: number;
  totalFeeExpected: number;
  totalFeeCollected: number;
}

export interface CredentialInfo {
  id: number;
  username: string;
  role: string;
  updatedAt: string;
}

export interface ChangePasswordInput {
  username: string;
  newPassword: string;
  currentPassword?: string;
  adminUsername?: string;
  adminPassword?: string;
}

export type ChildProgressFeeStatus = {
  currentMonth?: FeeRecord;
  pendingCount: number;
  totalDue: number;
};

export interface ChildProgress {
  student: Student;
  attendanceSummary: AttendanceSummary;
  feeStatus: ChildProgressFeeStatus;
  recentAttendance: AttendanceRecord[];
  announcements: Announcement[];
}

export type ListStudentsParams = {
class?: string;
search?: string;
whatsapp?: 'with' | 'without';
};

export type ListFeesParams = {
class?: string;
month?: number;
year?: number;
studentId?: number;
};

export type ListAttendanceParams = {
class?: string;
date?: string;
studentId?: number;
};

export type ListAnnouncementsParams = {
targetRole?: ListAnnouncementsTargetRole;
};

export type ListAnnouncementsTargetRole = typeof ListAnnouncementsTargetRole[keyof typeof ListAnnouncementsTargetRole];


export const ListAnnouncementsTargetRole = {
  all: 'all',
  admin: 'admin',
  parents: 'parents',
} as const;

export type GetDailyCollectionParams = {
date?: string;
};