# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Allied School Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /fees:
    get:
      operationId: listFees
      tags: [fees]
      summary: List fee records
      parameters:
        - name: class
          in: query
          schema:
            type: string
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
        - name: studentId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Fee records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeeRecord"
    post:
      operationId: createFeeRecord
      tags: [fees]
      summary: Create fee record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeeRecordInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeeRecord"
  /fees/{id}:
    patch:
      operationId: updateFeeRecord
      tags: [fees]
      summary: Update fee record (mark paid/unpaid)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeeRecordUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeeRecord"
  /fees/generate-monthly:
    post:
      operationId: generateMonthlyFees
      tags: [fees]
      summary: Generate monthly fee records for all students
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateMonthlyFeesInput"
      responses:
        "200":
          description: Generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List attendance records
      parameters:
        - name: class
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
        - name: studentId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
    post:
      operationId: markAttendance
      tags: [attendance]
      summary: Mark attendance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceInput"
      responses:
        "201":
          description: Marked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
  /attendance/{id}:
    patch:
      operationId: updateAttendance
      tags: [attendance]
      summary: Update attendance record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
  /announcements:
    get:
      operationId: listAnnouncements
      tags: [announcements]
      summary: List announcements
      parameters:
        - name: targetRole
          in: query
          schema:
            type: string
            enum: [all, admin, parents]
      responses:
        "200":
          description: Announcements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"
    post:
      operationId: createAnnouncement
      tags: [announcements]
      summary: Create announcement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /announcements/{id}:
    delete:
      operationId: deleteAnnouncement
      tags: [announcements]
      summary: Delete announcement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /settings/change-password:
    post:
      operationId: changePassword
      tags: [settings]
      summary: Change password for a user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    LoginInput:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    LoginResult:
      type: object
      properties:
        success:
          type: boolean
        role:
          type: string
          enum: [admin, director]
        username:
          type: string
      required: [success, role, username]
...
    FeeRecord:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        rollNumber:
          type: ["string", "null"]
        class:
          type: ["string", "null"]
        month:
          type: integer
        year:
          type: integer
        amount:
          type: integer
        paid:
          type: boolean
        paidDate:
          type: ["string", "null"]
      required: [id, studentId, month, year, amount, paid]
    FeeRecordInput:
      type: object
      properties:
        studentId:
          type: integer
        month:
          type: integer
        year:
          type: integer
        amount:
          type: integer
      required: [studentId, month, year, amount]
    FeeRecordUpdate:
      type: object
      properties:
        paid:
          type: boolean
        paidDate:
          type: string
    GenerateMonthlyFeesInput:
      type: object
      properties:
        month:
          type: integer
        year:
          type: integer
      required: [month, year]
    AttendanceRecord:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        rollNumber:
          type: ["string", "null"]
        class:
          type: ["string", "null"]
        date:
          type: string
        present:
          type: boolean
      required: [id, studentId, date, present]
    AttendanceInput:
      type: object
      properties:
        studentId:
          type: integer
        date:
          type: string
        present:
          type: boolean
      required: [studentId, date, present]
    AttendanceUpdate:
      type: object
      properties:
        present:
          type: boolean
...
    Announcement:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        targetRole:
          type: string
          enum: [all, admin, parents]
        createdAt:
          type: string
      required: [id, title, content, targetRole, createdAt]
    AnnouncementInput:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        targetRole:
          type: string
          enum: [all, admin, parents]
      required: [title, content, targetRole]
...
    ChangePasswordInput:
      type: object
      properties:
        username:
          type: string
        newPassword:
          type: string
      required: [username, newPassword]
...