# lib/api-zod/src/generated/types/feeRecord.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Allied School Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface FeeRecord {
  id: number;
  studentId: number;
  /** @nullable */
  studentName?: string | null;
  /** @nullable */
  rollNumber?: string | null;
  /** @nullable */
  class?: string | null;
  month: number;
  year: number;
  amount: number;
  paid: boolean;
  /** @nullable */
  paidDate?: string | null;
}