# lib/api-zod/src/generated/types/generateMonthlyFeesInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Allied School Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateMonthlyFeesInput {
  month: number;
  year: number;
}